Grammar Fixes and Clarification
Corrected all grammar
Major Changes:
Changed "How to Use" to "Usage Instructions"
Changed "Ctrl + Supp" to "Ctrl + Super(or Win) + Delete" as this was not
clear.
Changed various instances of "object" to "record" for user clarity.
Changed "accumulated extensive experience in project management in local
and Nearshore mode" to "gained extensive experience in project
management in local and Nearshore projects"